requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.4.2
pydantic-settings==2.0.3
httpx==0.25.2
authlib==1.2.1
python-dotenv==1.0.0
# Remove problematic packages for now
# redis==5.0.1
# celery==5.3.4
# Use psycopg2-binary instead of asyncpg for now
